+++ src/commands/AppCommands.js
@@ -7,11 +7,15 @@ const Enum = require('../Enum')
 const Meta = require('../Meta')
 
 const kebabCase = require('lodash.kebabcase')
+const pick = require('lodash.pick')
+const download = require('image-downloader')
 const fs = require('fs')
 const request = require('request')
 const path = require('path')
 const rp = require('request-promise')
 const asyncfile = require('async-file')
+const tempy = require('tempy')
+const compressing = require('compressing')
 
 class AppCommands {
 	static async register(appsProvider, _authorization, _environment, _admin) {
@@ -424,6 +428,278 @@ class AppCommands {
 					}
 					break
 			}
+		})
+
+		/**
+		 * Export app
+		 */
+		vscode.commands.registerCommand('apps-sdk.app.export', async function (context) {
+
+			// Context check
+			if (!Core.contextGuard(context)) { return }
+			let app = context
+
+			// Wait for a path for ZIP
+			let storage = await vscode.window.showSaveDialog({ filters: { 'IMT App Archive': ['zip'] }, saveLabel: "Export", defaultUri: vscode.Uri.file(`${app.name}.v${app.version}.zip`) })
+			if (storage === undefined || storage === null) {
+				vscode.window.showWarningMessage("No output file specified.")
+				return
+			}
+
+			await vscode.window.withProgress({
+				location: vscode.ProgressLocation.Notification,
+				title: `Exporting ${app.label}`,
+				cancellable: true
+			}, async (progress, token) => {
+
+				let canceled = false
+				progress.report({ increment: 0, message: `${app.label} - Preparing for export` })
+				const RATE_LIMIT_MS = 600
+				const DIR = tempy.directory()
+				token.onCancellationRequested(() => {
+					vscode.window.showWarningMessage(`Export of ${app.label} canceled.`);
+					canceled = true;
+				})
+
+				let progressPercentage;
+
+				const archive = path.join(DIR, app.name)
+				await asyncfile.mkdir(archive);
+				const urnNoVersion = `${_environment}/app/${app.name}`
+				const urn = `${_environment}/app/${app.name}/${app.version}`
+
+				try {
+					/**
+					 * 1 - Get App Metadata
+					 */
+					if (canceled) { return }
+					progress.report({ increment: 2, message: `${app.label} - Gathering Metadata` })
+					await asyncfile.writeFile(path.join(archive, `metadata.json`), JSON.stringify(pick(await Core.rpGet(`${urn}`, _authorization), ['name', 'label', 'version', 'theme', 'language', 'countries']), null, 4));
+					await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+					/**
+					 * 2 - Get Base
+					 */
+					if (canceled) { return }
+					progress.report({ increment: 2, message: `${app.label} - Exporting Base` })
+					await asyncfile.writeFile(path.join(archive, `base.imljson`), Core.jsonString(await Core.rpGet(`${urn}/base`, _authorization)))
+					await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+					/**
+					 * 3 - Get Readme
+					 */
+					if (canceled) { return }
+					progress.report({ increment: 2, message: `${app.label} - Exporting Readme` })
+					await asyncfile.writeFile(path.join(archive, `readme.md`), await Core.rpGet(`${urn}/readme`, _authorization))
+					await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+					/**
+					 * 4 - Get Connections
+					 */
+					if (canceled) { return }
+					const connections = await Core.rpGet(`${urnNoVersion}/connection`, _authorization);
+					if (connections.length === 0) { progress.report({ increment: 7, message: `${app.label} - No Connections (skipping)` }) }
+					else { progressPercentage = 7 / connections.length }
+					await asyncfile.mkdir(path.join(archive, 'connection'))
+					for (const connection of connections) {
+						if (canceled) { return }
+						const archivePath = path.join(archive, 'connection', connection.name)
+						await asyncfile.mkdir(archivePath)
+
+						// Get Connection Metadata
+						progress.report({ increment: 0.125 * progressPercentage, message: `${app.label} - Exporting Connection ${connection.label} (metadata)` })
+						await asyncfile.writeFile(path.join(archivePath, `metadata.json`), JSON.stringify(pick(await Core.rpGet(`${urnNoVersion}/connection/${connection.name}`, _authorization), ['name', 'label', 'type']), null, 4));
+						await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+						// Get Corresponding Sources
+						for (const key of [`api`, `scope`, `scopes`, `parameters`]) {
+							progress.report({ increment: (0.875 * progressPercentage) * (0.25), message: `${app.label} - Exporting Connection ${connection.label} (${key})` })
+							await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urnNoVersion}/connection/${connection.name}/${key}`, _authorization)))
+							await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+						}
+					}
+
+					/**
+					 * 5 - Get RPCs
+					 */
+					if (canceled) { return }
+					const rpcs = await Core.rpGet(`${urn}/rpc`, _authorization);
+					if (rpcs.length === 0) { progress.report({ increment: 17, message: `${app.label} - No RPCs (skipping)` }) }
+					else { progressPercentage = 17 / rpcs.length }
+					await asyncfile.mkdir(path.join(archive, 'rpc'))
+					for (const rpc of rpcs) {
+						if (canceled) { return }
+						const archivePath = path.join(archive, 'rpc', rpc.name)
+						await asyncfile.mkdir(archivePath)
+
+						// Get RPC Metadata
+						progress.report({ increment: (0.25 * progressPercentage), message: `${app.label} - Exporting RPC ${rpc.label} (metadata)` })
+						await asyncfile.writeFile(path.join(archivePath, `metadata.json`), JSON.stringify(pick(await Core.rpGet(`${urn}/rpc/${rpc.name}`, _authorization), ['name', 'label', 'connection']), null, 4));
+						await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+						// Get Corresponding Sources
+						for (const key of [`api`, `parameters`]) {
+							progress.report({ increment: (0.75 * progressPercentage) * (0.5), message: `${app.label} - Exporting RPC ${rpc.label} (${key})` })
+							await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urn}/rpc/${rpc.name}/${key}`, _authorization)))
+							await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+						}
+					}
+
+					/**
+					 * 6 - Get Webhooks
+					 */
+					if (canceled) { return }
+					const webhooks = await Core.rpGet(`${urnNoVersion}/webhook`, _authorization);
+					if (webhooks.length === 0) { progress.report({ increment: 12, message: `${app.label} - No Webhooks (skipping)` }) }
+					else { progressPercentage = 12 / webhooks.length }
+					await asyncfile.mkdir(path.join(archive, 'webhook'))
+					for (const webhook of webhooks) {
+						if (canceled) { return }
+						const archivePath = path.join(archive, 'webhook', webhook.name)
+						await asyncfile.mkdir(archivePath)
+
+						// Get Webhook Metadata
+						progress.report({ increment: 0.1 * progressPercentage, message: `${app.label} - Exporting Webhook ${webhook.label} (metadata)` })
+						await asyncfile.writeFile(path.join(archivePath, `metadata.json`), JSON.stringify(pick(await Core.rpGet(`${urnNoVersion}/webhook/${webhook.name}`, _authorization), ['name', 'label', 'connection', 'type']), null, 4));
+						await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+						// Get Corresponding Sources
+						for (const key of [`api`, `parameters`, `attach`, `detach`, `scope`]) {
+							progress.report({ increment: (0.9 * progressPercentage) * (0.2), message: `${app.label} - Exporting Webhook ${webhook.label} (${key})` })
+							await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urnNoVersion}/webhook/${webhook.name}/${key}`, _authorization)))
+							await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+						}
+					}
+
+					/**
+					 * 7 - Get Modules
+					 */
+					if (canceled) { return }
+					const modules = await Core.rpGet(`${urn}/module`, _authorization);
+					if (modules.length === 0) { progress.report({ increment: 41, message: `${app.label} - No Modules (skipping)` }) }
+					else { progressPercentage = 41 / modules.length }
+					await asyncfile.mkdir(path.join(archive, 'module'))
+					for (const module of modules) {
+						if (canceled) { return }
+						const archivePath = path.join(archive, 'module', module.name)
+						await asyncfile.mkdir(archivePath)
+
+						// Get Module Metadata
+						progress.report({ increment: 0.07 * progressPercentage, message: `${app.label} - Exporting Module ${module.label} (metadata)` })
+						let metadata = pick(await Core.rpGet(`${urn}/module/${module.name}`, _authorization), ['name', 'label', 'description', 'type_id', 'connection', 'webhook'])
+						switch (metadata.type_id) {
+							case 1:
+								metadata.type = 'trigger'
+								break;
+							case 4:
+								metadata.type = 'action'
+								break;
+							case 9:
+								metadata.type = 'search'
+								break;
+							case 10:
+								metadata.type = 'instant_trigger'
+								break;
+							case 11:
+								metadata.type = 'responder'
+								break;
+						}
+						delete metadata.type_id
+						await asyncfile.writeFile(path.join(archivePath, `metadata.json`), JSON.stringify(metadata, null, 4));
+						await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+
+						// Get Corresponding Sources Based On Type
+						switch (module.type_id) {
+
+							// Action or search
+							case 4:
+							case 9:
+								for (const key of [`api`, `parameters`, `expect`, `interface`, `samples`, `scope`]) {
+									progress.report({ increment: (0.93 * progressPercentage) * (0.16), message: `${app.label} - Exporting Module ${module.label} (${key})` })
+									await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urn}/module/${module.name}/${key}`, _authorization)))
+									await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+								}
+								break;
+							// Trigger
+							case 1:
+								for (const key of [`api`, `epoch`, `parameters`, `interface`, `samples`, `scope`]) {
+									progress.report({ increment: (0.93 * progressPercentage) * (0.16), message: `${app.label} - Exporting Module ${module.label} (${key})` })
+									await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urn}/module/${module.name}/${key}`, _authorization)))
+									await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+								}
+								break;
+							// Instant trigger
+							case 10:
+								for (const key of [`api`, `parameters`, `interface`, `samples`]) {
+									progress.report({ increment: (0.93 * progressPercentage) * (0.25), message: `${app.label} - Exporting Module ${module.label} (${key})` })
+									await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urn}/module/${module.name}/${key}`, _authorization)))
+									await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+								}
+								break;
+							// Responder
+							case 11:
+								for (const key of [`api`, `parameters`, `expect`]) {
+									progress.report({ increment: (0.93 * progressPercentage) * (0.33), message: `${app.label} - Exporting Module ${module.label} (${key})` })
+									await asyncfile.writeFile(path.join(archivePath, `${key}.imljson`), Core.jsonString(await Core.rpGet(`${urn}/module/${module.name}/${key}`, _authorization)))
+									await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+								}
+								break;
+						}
+					}
+
+					/**
+					 * 8 - Get Functions
+					 */
+					if (canceled) { return }
+					const functions = await Core.rpGet(`${urn}/function`, _authorization);
+					if (functions.length === 0) { progress.report({ increment: 7, message: `${app.label} - No Functions (skipping)` }) }
+					else { progressPercentage = 7 / functions.length }
+					await asyncfile.mkdir(path.join(archive, 'function'))
+					for (const fun of functions) {
+						if (canceled) { return }
+						const archivePath = path.join(archive, 'function', fun.name)
+						await asyncfile.mkdir(archivePath)
+
+						// Get Corresponding Sources
+						for (const key of [`code`, `test`]) {
+							progress.report({ increment: progressPercentage * 0.5, message: `${app.label} - Exporting Function ${fun.name}${fun.args} (${key})` })
+							await asyncfile.writeFile(path.join(archivePath, `${key}.js`), await Core.rpGet(`${urn}/function/${fun.name}/${key}`, _authorization))
+							await new Promise(resolve => setTimeout(resolve, RATE_LIMIT_MS));
+						}
+					}
+
+					/**
+					 * 9 - Get Icon
+					 */
+					if (canceled) { return }
+					progress.report({ increment: 3, message: `${app.label} - Exporting Icon` })
+					await asyncfile.mkdir(path.join(archive, 'assets'))
+					try {
+						await download.image({
+							headers: {
+								"Authorization": _authorization,
+								'x-imt-apps-sdk-version': Meta.version
+							},
+							url: `${urn}/icon/512`,
+							dest: path.join(archive, 'assets', 'icon.png')
+						})
+					}
+					catch (err) { }
+
+					/**
+					 * 10 - Compress and save
+					 */
+					if (canceled) { return }
+					progress.report({ increment: 5, message: `${app.label} - Compressing output` })
+					await compressing.zip.compressDir(archive, storage.fsPath).catch((err) => {
+						vscode.window.showErrorMessage(`Saving failed: ${err}`)
+					})
+				}
+				catch (err) {
+					vscode.window.showErrorMessage(err)
+				}
+			});
+			vscode.window.showInformationMessage(`Export of ${app.label} completed!`)
 		})
 	}
 }