+++ src/commands/CoreCommands.js
@@ -7,6 +7,7 @@ const RpcProvider = require('../providers/RpcProvider')
 const ImlProvider = require('../providers/ImlProvider')
 const ParametersProvider = require('../providers/ParametersProvider')
 const StaticImlProvider = require('../providers/StaticImlProvider')
+const TempProvider = require('../providers/TempProvider')
 
 const path = require('path')
 const fs = require('fs')
@@ -15,7 +16,7 @@ const mkdirp = require('mkdirp')
 const request = require('request')
 
 class CoreCommands {
-	constructor(appsProvider, _authorization, _environment, rpcProvider, imlProvider, parametersProvider, staticImlProvider) {
+	constructor(appsProvider, _authorization, _environment, rpcProvider, imlProvider, parametersProvider, staticImlProvider, tempProvider) {
 		this.appsProvider = appsProvider
 		this._authorization = _authorization
 		this._environment = _environment
@@ -23,8 +24,10 @@ class CoreCommands {
 		this.currentImlProvider = imlProvider
 		this.currentParametersProvider = parametersProvider
 		this.currentStaticImlProvider = staticImlProvider
+		this.currentTempProvider = tempProvider
 		this.staticImlProvider = new StaticImlProvider()
 		this.sipInit = false
+		this.tempListener = null
 	}
 
     /**
@@ -148,8 +151,11 @@ class CoreCommands {
 
 			// If no version, set 1
 			version = version === undefined ? 1 : version;
+			let apiPath
+			let name
+			let app = crumbs[2]
 
-			// If not enough crumbs (=> base/docs) -> remove existing providers and exit.
+			// If not enough crumbs (=> base/readme) -> remove existing providers and exit.
 			if (crumbs.length < 5) {
 
 				// Remove existing RpcProvider
@@ -157,28 +163,38 @@ class CoreCommands {
 					this.currentRpcProvider.dispose()
 				}
 
-				// Remove existing ImlProvider
-				if (this.currentImlProvider !== null && this.currentImlProvider !== undefined) {
-					this.currentImlProvider.dispose()
-				}
-
 				// Remove existing ParametersProvider
 				if (this.currentParametersProvider !== null && this.currentParametersProvider !== undefined) {
 					this.currentParametersProvider.dispose()
 				}
 
-				// Remove existing StaticImlProvider
-				if (this.currentStaticImlProvider !== null && this.currentStaticImlProvider !== undefined) {
-					this.currentStaticImlProvider.dispose()
+				if (crumbs[3] !== 'base.imljson') {
+					// Remove existing ImlProvider
+					if (this.currentImlProvider !== null && this.currentImlProvider !== undefined) {
+						this.currentImlProvider.dispose()
+					}
+
+					// Remove existing StaticImlProvider
+					if (this.currentStaticImlProvider !== null && this.currentStaticImlProvider !== undefined) {
+						this.currentStaticImlProvider.dispose()
+					}
+
+					// Remove existing TempProvider
+					if (this.currentTempProvider !== null && this.currentTempProvider !== undefined) {
+						this.currentTempProvider.dispose()
+					}
+
+					return
 				}
 
-				return
+				apiPath = 'base'
+				name = 'base'
 			}
 
-			let apiPath = crumbs[3]
-			let name = crumbs[5].split(".")[0]
-			let app = crumbs[2]
-
+			else {
+				apiPath = crumbs[3]
+				name = crumbs[5].split(".")[0]
+			}
 
             /**
              * RPC-LOADER
@@ -225,7 +241,7 @@ class CoreCommands {
              * Following condition specifies where are IML functions allowed
              */
 			if (
-				(name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
+				(name === "base" || name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
 			) {
 				// Remove existing ImlProvider
 				if (this.currentImlProvider !== null && this.currentImlProvider !== undefined) {
@@ -262,7 +278,7 @@ class CoreCommands {
              */
 
 			if (
-				(name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
+				apiPath !== 'base' && (name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
 			) {
 				// Remove existing ParametersProvider
 				if (this.currentParametersProvider !== null && this.currentParametersProvider !== undefined) {
@@ -297,7 +313,7 @@ class CoreCommands {
 			}
 
 			if (
-				(name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
+				(name === "base" || name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
 			) {
 				// Reasign Static IML provider
 				if (this.currentStaticImlProvider !== null && this.currentStaticImlProvider !== undefined) {
@@ -314,6 +330,33 @@ class CoreCommands {
 				// If out of scope, remove existing StaticImlProvider
 				if (this.currentStaticImlProvider !== null && this.currentStaticImlProvider !== undefined) {
 					this.currentStaticImlProvider.dispose()
+				}
+			}
+
+			/**
+			 * TEMP-LOADER
+			 * Temp loader provides available temp variables in the document
+			 */
+
+			if (
+				(name === "base" || name === "api" || name === "api-oauth" || name === "epoch" || name === "attach" || name === "detach")
+			) {
+				// Reasign Temp provider
+				if (this.currentTempProvider !== null && this.currentTempProvider !== undefined) {
+					this.currentTempProvider.dispose()
+				}
+
+				let tempProvider = new TempProvider(name);
+
+				this.currentTempProvider = vscode.languages.registerCompletionItemProvider({
+					scheme: 'file',
+					language: 'imljson'
+				}, tempProvider)
+			}
+			else {
+				// If out of scope, remove existing StaticImlProvider
+				if (this.currentTempProvider !== null && this.currentTempProvider !== undefined) {
+					this.currentTempProvider.dispose()
 				}
 			}
 		}
@@ -348,8 +391,8 @@ class CoreCommands {
 				case "app":
 					// Prepared for more app-level codes
 					switch (item.name) {
-						case "content":
-							urn += `/docs`
+						case "readme":
+							urn += `/readme`
 							break
 						default:
 							urn += `/${item.name}`
@@ -440,7 +483,7 @@ class CoreCommands {
 					// Prepared for more app-level codes
 					switch (item.name) {
 						case "content":
-							urn += `/docs`
+							urn += `/readme`
 							break
 						default:
 							urn += `/${item.name}`